wspexam/requirements.txt
Django==2.2
django-bootstrap3==11.0.0
django-crispy-forms==1.7.2
django-model-utils==3.1.2
examapp==0.0.5
pytz==2018.9
sqlparse==0.3.0
